SW pipeline: fixing division by ZERO (RT-28171, RT-27905 ?)
- remove JNI calls out of Get/ReleasePrimitiveArrayCritical block
- check for NULL value returned by GetPrimitiveArrayCritical

+++ prism-sw-native/include/PiscesRenderer.h
@@ -321,7 +321,7 @@ typedef struct _Renderer {
     
     // Paint buffer
     jint *_paint;
-    jint _paint_length;
+    size_t _paint_length;
 
     // Paint transform
     Transform6 _paint_transform;
